feat: change /image endpoint from GET to POST
This commit changes the /image endpoint from a GET request to a POST request.
It also adds a new 'type' parameter to the request body, which allows the client
to specify the format of the image they want to receive. The response format
has been updated to accommodate this change. If the 'type' is 'base64', the
image is returned as a JSON object. Otherwise, the image is sent directly with
the appropriate 'Content-Type' header.